apps/crawler/requirements.txt
# Packman Crawler
# Discovers MCU libraries from GitHub, scores quality, and writes to the DB.

PyGithub>=2.1.1
httpx>=0.27.0
apscheduler>=3.10.4
psycopg[binary]>=3.1.18
python-dotenv>=1.0.1
pydantic>=2.6.0
pyyaml>=6.0.2
toml>=0.10.2
